del report and move get_employee_tasks into process
