+++ project1_task-management/solution/main.py
@@ -22,6 +22,8 @@ Functions:
 - MOO(model, employees, company_tasks, score, story_points, max_employee_workload, mu_Z_1, mu_Z_2, mu_Z_3, mu_Z_star, assessment_score_1, assessment_score_2, assessment_score_3): Multi-Objective Optimization using Goal Programming.
 
 Classes:
+- CompetencyAssessment: A class to assess the competencies of employees against required competencies for tasks using MSG scores.
+- WeightedEuclideanDistance: A class to assess the competencies of employees against required competencies for tasks using Weighted Euclidean Distance.
 - GapCallback: A callback class to report optimization progress and gap.
 
 Usage:
@@ -50,7 +52,7 @@ from optimizer import (
     creds,
     config,
     helper,
-    preprocessing,
+    process,
     create_model,
     create_objective,
 )
@@ -66,7 +68,7 @@ def main():
     try:
         # Section 1
         employees, tasks, story_points, company_tasks, score, info = (
-            preprocessing.define_data_structure(
+            process.define_data_structure(
                 creds.employee_path, creds.task_path, config.overqualification
             )
         )
@@ -231,7 +233,7 @@ def main():
             "MOO with\nGoal Programming",
         ]
 
-        preprocessing.compare_scores(data, title, "output/score_comaprison")
+        process.compare_scores(data, title, "output/score_comaprison")
 
     except Exception as e:
         helper.send_discord_notification(f"An error occurred: {e}")
