+++ project1_task-management/solution/optimizer/create_objective.py
@@ -1,10 +1,9 @@
-import threading
 import pandas as pd
 import matplotlib.pyplot as plt
 from gurobipy import GRB, Model, quicksum
 from typing import Dict, List, Tuple, Any
 from optimizer.callback import GapCallback
-from . import report, helper, config, preprocessing
+from . import helper, config, process
 
 
 def objective1(
@@ -56,7 +55,7 @@ def objective1(
 
             x_hat_1 = {}
             for j in employees:
-                result = report.get_employee_tasks(
+                result = process.get_employee_tasks(
                     j, company_tasks, model, score, story_points, max_employee_workload
                 )
                 if len(result[1]) > 0:
@@ -66,7 +65,7 @@ def objective1(
             x_hat_1 = {}
 
         # Call the process_results function
-        assessment_score_1 = preprocessing.process_results(
+        assessment_score_1 = process.process_results(
             x_hat_1,
             employees,
             story_points,
@@ -131,7 +130,7 @@ def objective2(
 
             x_hat_2 = {}
             for j in employees:
-                result = report.get_employee_tasks(
+                result = process.get_employee_tasks(
                     j, company_tasks, model, score, story_points, max_employee_workload
                 )
                 if len(result[1]) > 0:
@@ -141,7 +140,7 @@ def objective2(
             x_hat_2 = {}
 
         # Call the process_results function
-        assessment_score_2 = preprocessing.process_results(
+        assessment_score_2 = process.process_results(
             x_hat_2,
             employees,
             story_points,
@@ -203,7 +202,7 @@ def objective3(
 
             x_hat_3 = {}
             for j in employees:
-                result = report.get_employee_tasks(
+                result = process.get_employee_tasks(
                     j, company_tasks, model, score, story_points, max_employee_workload
                 )
                 if len(result[1]) > 0:
@@ -213,7 +212,7 @@ def objective3(
             x_hat_3 = {}
 
         # Call the process_results function
-        assessment_score_3 = preprocessing.process_results(
+        assessment_score_3 = process.process_results(
             x_hat_3,
             employees,
             story_points,
@@ -321,7 +320,7 @@ def MOO(
 
             x_hat_4 = {}
             for j in employees:
-                result = report.get_employee_tasks(
+                result = process.get_employee_tasks(
                     j, company_tasks, model, score, story_points, max_employee_workload
                 )
                 if len(result[1]) > 0:
@@ -331,7 +330,7 @@ def MOO(
             x_hat_4 = {}
 
         # Call the process_results function
-        assessment_score_4 = preprocessing.process_results(
+        assessment_score_4 = process.process_results(
             x_hat_4,
             employees,
             story_points,
