+++ project1_task-management/solution/optimizer/tools.py
@@ -66,7 +66,6 @@ Last Modified:
 June 2024
 """
 
-import math
 import heapq
 import numpy as np
 import pandas as pd
